安定稼働版

+++ database/migrations/2017_09_29_092246_create_chochiku_account_ledgers.php
@@ -43,7 +43,7 @@ class CreateChochikuAccountLedgers extends Migration
             $table->char("spare_1", 27);
             $table->integer("monthly_id")->index();
             $table->integer("subject_code")->index();
-            $table->integer("key_account_number")->index();
+            $table->double("key_account_number")->index();
             $table->timestamps("");
         });
     }
